+++ app/src/main/java/com/iosharp/android/ssplayer/EventListFragment.java
@@ -42,6 +42,8 @@ public class EventListFragment extends Fragment {
     }
 
     private static void getDateEvents(Context context, ArrayList<String> dates, ArrayList<ArrayList<Event>> events) {
+
+
         Uri uri = EventEntry.buildEventDate();
         Cursor dateCursor = context.getContentResolver().query(uri, null, null, null, null);
 
@@ -89,6 +91,8 @@ public class EventListFragment extends Fragment {
     }
 
     public static void updateEvents(Context context) {
+        mDate.clear();
+        mDateEvents.clear();
 
         getDateEvents(context, mDate, mDateEvents);
         mAdapter.notifyDataSetChanged();